refactor: 모달 클릭, 키 입력 이벤트 훅 분리, 스크롤 이벤트 방지 기능 추가
